+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 Released on 30/09/2024
 
 - Moved project to remotefs-rs org
+- StreamReader and StreamWriter implement Send
 
 ## 0.2.0