+++ CHANGELOG.md
@@ -1,12 +1,20 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.3.1](#031)
   - [0.3.0](#030)
   - [0.2.0](#020)
   - [0.1.1](#011)
   - [0.1.0](#010)
 
 ---
+
+## 0.3.1
+
+Released on 22/10/2024
+
+- `FileType` is now `Copy`
+- `stream` in `StreamWriter` is now `pub`
 
 ## 0.3.0